genetic_indicator_optimization/requirements.txt
# Основные библиотеки для работы с данными
pandas>=2.0.0
numpy>=1.24.0

# Визуализация
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Генетические алгоритмы
deap>=1.4.0

# Технический анализ
ta-lib>=0.4.0
pandas-ta>=0.3.14b

# Машинное обучение и метрики
scikit-learn>=1.3.0
scipy>=1.10.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Конфигурация
pyyaml>=6.0

# Утилиты
tqdm>=4.65.0
joblib>=1.3.0

# Оптимизация производительности
numba>=0.57.0

